Almost all warnings fixed

+++ GWToolboxdll/GWToolbox/Widgets/Minimap/RangeRenderer.h
@@ -7,9 +7,9 @@
 
 class RangeRenderer : public VBuffer {
 private:
-	static const int num_circles = 5;
-	static const int circle_points = 64;
-	static const int circle_vertices = 65;
+	static const size_t num_circles = 5;
+	static const size_t circle_points = 64;
+	static const size_t circle_vertices = 65;
 
 public:
 	void Render(IDirect3DDevice9* device) override;
